Project3/Project-3/requirements.txt
# ─── Core Agent Framework ───────────────────────────────────
crewai
crewai-tools

# ─── LLM Providers ──────────────────────────────────────────
google-generativeai
groq
ollama
langchain-google-genai
langchain-groq
langchain-community
langchain-core

# ─── Monitoring ─────────────────────────────────────────────
langfuse

# ─── UI ─────────────────────────────────────────────────────
streamlit

# ─── Utilities ──────────────────────────────────────────────
python-dotenv
pyyaml
rich